requirements.txt
faiss-cpu
pdfplumber
python-docx
tavily
sentence-transformers
openai
python-dotenv
fastapi
uvicorn
pydantic